refactor(stella-tui): move the recall read-model into model/recall.rs
`make gate` caught two file-size violations from the previous commit, and
both take the fix the guard asks for rather than a baseline entry.

`model.rs` crossed the 1500-line limit (1449 → 1571). It is not
grandfathered, so the rule is explicit: split into a sibling submodule,
and do not add a baseline entry — that list only covers files predating
the guard. `RecalledFrameRow`, `RecallBudget` and the wire projection move
to `model/recall.rs`, beside the existing `model/file_state.rs`. The cut
follows the concern: everything moved is about one event,
`AgentEvent::ContextRecall`, and nothing else in the fold reads it. The
fold arm shrinks to a destructuring call, which also makes the
"lose nothing at this seam" rule enforceable in one place — that seam
narrowing is why `latency_ms` and `used_ann_index` never reached a
surface. `model.rs` lands at 1488.

`deck_ui.rs` grew 4006 → 4013 on a doc comment. It is a grandfathered god
file and closed to growth, so the comment comes back out; the paragraph
it carried — that `is_expandable` and the `expanded` arms of
`render::entry_lines` are one contract in two places that had already
drifted — moves to `render/entry.rs`, which is not a god file and holds
the other half of that contract. `deck_ui.rs` is back at exactly 4006.

No behaviour change: `cargo test -p stella-tui` is 836 passed, and
`make gate CARGO_SCOPE="-p stella-tui"` is green across all 26 steps.
