+++ crates/stella-tui/src/model.rs
@@ -28,9 +28,11 @@ use std::collections::VecDeque;
 
 mod error_rows;
 pub mod file_state;
+pub mod recall;
 #[cfg(test)]
 pub use file_state::DIFF_HISTORY;
 pub use file_state::{FileState, MAX_TRACKED_FILES, RememberedDiff};
+pub use recall::{RecallBudget, RecalledFrameRow};
 
 /// How many characters of a tool input / output summary we retain on a
 /// transcript line before eliding — the full payload is never needed on the
@@ -207,62 +209,6 @@ pub struct SubAgentSummary {
     pub absorbed_messages: usize,
     /// Why it did not complete; `None` when it did.
     pub reason: Option<String>,
-}
-
-/// One recalled context frame, as [`TranscriptEntry::ContextRecall`] holds it.
-///
-/// A near-lossless read-model of `stella_protocol::ContextFrameRef`: every
-/// field the wire carries survives except `block_id`, which recall never sets
-/// (a frame becomes a block only once it is *rendered* into a message, after
-/// the event is emitted — see `Recall::telemetry_event`).
-///
-/// The two fields worth naming are the two the old label-only shape lost.
-/// [`Self::kind`] is what separates a 60-token graph symbol from an 800-token
-/// episodic memory, and until it survived to the renderer the deck showed both
-/// as undifferentiated prose. [`Self::tokens`] is what turns a recall total
-/// into a finding: "1155 tok" is a number, but one frame holding 812 of them
-/// is a reason to tune retrieval.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct RecalledFrameRow {
-    /// Protocol frame kind — `symbol`, `memory`, `episode`, `fact`,
-    /// `snippet`, `doc`, `graph`. Empty on a stream recorded before the field
-    /// existed, which the renderer shows as `frame` rather than a blank column.
-    pub kind: String,
-    /// The human citation (L-C4) — never a raw id.
-    pub label: String,
-    /// The frame's canonical source URI, when it declared one.
-    pub uri: Option<String>,
-    /// The CGP provider leg that served the frame (`code-graph`,
-    /// `workspace-memory`).
-    pub provider: String,
-    /// The original source named by the provenance chain — deliberately
-    /// distinct from [`Self::provider`], which is the adapter that fronted it.
-    pub source: String,
-    /// The most-derived provenance method, when declared.
-    pub method: Option<String>,
-    /// The provider's own frame id, a detail-view identifier only (L-C4).
-    pub id: Option<String>,
-    /// `"sha256:<hex>"` over the exact injected bytes. Its **absence** is
-    /// meaningful — such a frame is not verifiable per the context-reuse spec
-    /// — so the expanded view says "unverifiable" rather than showing nothing.
-    pub digest: Option<String>,
-    /// What injecting this frame cost the prompt.
-    pub tokens: u32,
-}
-
-/// The CGP budget report for one recall, as the deck's expanded view shows it.
-///
-/// `served`/`rejected` is the pair that earns this its place: a provider that
-/// served five frames and had two rejected has a cost-honesty problem the
-/// frame list alone cannot show, because rejected frames never reach it.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct RecallBudget {
-    /// Tokens the host asked for.
-    pub requested: u64,
-    /// Tokens the recall actually spent.
-    pub consumed: u64,
-    /// `(provider, served, rejected, tokens)`, one per contributing leg.
-    pub providers: Vec<(String, u32, u32, u64)>,
 }
 
 /// One semantic entry in the transcript. Rendering (colour, borders, glyphs)
@@ -874,44 +820,15 @@ impl SessionModel {
                 latency_ms,
                 used_ann_index,
             } => {
+                let (frames, providers, budget) =
+                    recall::project(frames, provider_mix, usage.as_ref());
                 self.transcript.push(TranscriptEntry::ContextRecall {
-                    frames: frames
-                        .iter()
-                        .map(|f| RecalledFrameRow {
-                            kind: f.kind.clone(),
-                            label: f.citation_label.clone(),
-                            uri: f.uri.clone(),
-                            provider: f.provider.clone(),
-                            source: f.source.clone(),
-                            method: f.method.clone(),
-                            id: f.id.clone(),
-                            digest: f.content_digest.clone(),
-                            tokens: f.token_cost,
-                        })
-                        .collect(),
+                    frames,
                     tokens: *tokens,
                     latency_ms: *latency_ms,
                     used_ann_index: *used_ann_index,
-                    providers: provider_mix
-                        .iter()
-                        .map(|s| (s.provider.clone(), s.frames))
-                        .collect(),
-                    budget: usage.as_ref().map(|u| RecallBudget {
-                        requested: u64::from(u.budget_requested),
-                        consumed: u.budget_consumed,
-                        providers: u
-                            .providers
-                            .iter()
-                            .map(|p| {
-                                (
-                                    p.provider_id.clone(),
-                                    p.frames_served,
-                                    p.frames_rejected,
-                                    p.token_cost,
-                                )
-                            })
-                            .collect(),
-                    }),
+                    providers,
+                    budget,
                 });
             }
             AgentEvent::ContextWrite {
